+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 
 * feature#497: Show acknowledgment status on Thresholds tab instead of log
 
+* feature#500: Re-warning cycle when alert is reached
+
 * feature#528: Allow Thold to Use the BCC in addition to the TO line for Notification Lists
 
 * feature#569: Provide enough data in environment of command to get Graph Information about Thold